+++ src/config.py
@@ -2,7 +2,7 @@ import os
 
 #  The bot will declare its version in the welcome message.
 # Updated by the release pipeline. Change manually if building from source
-BOT_VERSION = "v.local"
+BOT_VERSION = "v2.0.5"
 # Enable verbose application diagnostics in both the file and container logs.
 DEBUG = os.getenv("DEBUG", "false").strip().lower() in ["true", "1", "yes", "on"]
 # Add your stuff here